backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
openai==1.30.1
pdfplumber==0.11.0
python-docx==1.1.2
fpdf2==2.7.9
serpapi==0.1.5
google-search-results==2.4.2
aiofiles==23.2.1
passlib[bcrypt]==1.7.4
cryptography==42.0.7
pytest==8.2.1
pytest-asyncio==0.23.7
httpx==0.27.0
anyio==4.3.0
